+++ src/Components/HeaderComponent/TopPromotion.js
@@ -1,0 +1,19 @@
+import React, { useState} from "react";
+import { useEffect } from 'react';
+
+
+function TopPromotion() {
+  
+  
+  return (
+    <div className="header-container">
+      <div className="promotion-bar-top">
+        <div className="container mx-auto text-center padding-top-left-5px text-white">
+          <p className="text-14 mx-auto-0">Valentine promotion available*</p>
+        </div>
+      </div>
+
+    </div>
+  );
+}
+export default TopPromotion;
