+++ src/css/App.css
@@ -133,9 +133,9 @@ p {
 .search-bar {
   width: 60%;
   display: flex;
-
+  
   border-style: solid;
-  margin: 1.2rem;
+  margin: 0.5rem;
   overflow: hidden;
   /* test */
   transform: scale(0.9);
@@ -278,6 +278,9 @@ p {
 .width-100-px {
   width: 100px;
 }
+.width-150-px {
+  width: 150px;
+}
 .height-100{
   height: 100%;
 }
@@ -309,7 +312,7 @@ p {
   position: block;
   background-color: #f1f1f1;
   padding-top: 1rem !important;
-  
+  border-radius: 5px;
   
   width: 100%;
   box-shadow: 0px 8px 16px 0px rgba(0,0,0,0.2);
@@ -324,6 +327,7 @@ p {
   display: inline-block;
   background-image: linear-gradient(90deg, #d31e4b, #93188f);
   color: white;
+  border-radius: 8px;
 }
 
 .category-hoverable:hover > .category-hoverable-content{
@@ -331,6 +335,7 @@ p {
   background-image: linear-gradient(90deg, #d31e4b, #93188f);
   color: white;
 }
+
 .sub-menu-cover{
   padding-top: 0.8rem;
   padding-bottom: 0.8rem;
